Cleaned up team controller file
